feat: sanitize filename
